[ASHF-35] Update ma-dev (#21)
* Error strings should not be capitalized (unless beginning with proper nouns or acronyms) or end with punctuation

Signed-off-by: wuyingjun <wuyingjun@cmss.chinamobile.com>

* [FAB-17778] MSP.IsWellFormed: Only allow canonical CA signatures

This change makes MSP.IsWellFormed (not used in transaction validation, only for gossip)
for x509 based identities to check that the signature of the CA over an identity has canonical ASN1 encoding.
It also makes sure that the certificate is encoded in a single and minimal PEM block.

Change-Id: I7ba0db00b3eb58171bb05ee7680721a41ef1c802
Signed-off-by: yacovm <yacovm@il.ibm.com>

* Errors should be check and handle when initFactories is called

Signed-off-by: wuyingjun <wuyingjun@cmss.chinamobile.com>

* make couchdb index processor deterministic (#1121)

We need to satisfy two requirements while processing the index files.
R1: all valid indexes should be processed.
R2: the order of index creation must be the same in all peers. For example, if user
passes two index files with the same index name but different index fields and we
process these files in different orders in different peers, each peer would
have different indexes (as one index definion could replace another if the index names
are the same).

To satisfy R1, we log the error and continue to process the next index file.
To satisfy R2, we sort the indexFilesData map based on the filenames and process
each index as per the sorted order.

Signed-off-by: senthil <cendhu@gmail.com>

* rm unnecessary mock from txmgr pkg (#1183)

In the mvcc validator, we already have a mock for TxMgr
interface but use the mock located at txmgr package instead.
Hence, we make the validator to use the mock that present
locally and remove the one located in the txmgr package.

Signed-off-by: senthil <cendhu@gmail.com>

* cp config struct for testing couchdb failure scenarios (#1184)

We make changes to retries and address fields in config
struct to test some failure scenarios. To ensure proper cleanup
of the test environment, we revert those changes. It
looks hacky and hence, we want to copy the config and use
it for testing failure scenarios.

Signed-off-by: senthil <cendhu@gmail.com>

* [LEDGER] ensure peer panic does not occur when the index path is wrong (#1181)

* check len. of indexDir before accessing

While processing the couchdb index files, we retrieve and store
all directories in the index filepath into an array in order to
find whether it is a chaincode index or collection index. Currently,
we access the array without checking the number of directories
in the path assuming that the user would pass the correct path.
When the user passes a wrong path, the current array access would
result in a peer panic. Hence, we add a check on the length of
the array before accessing it.

FAB-17771

Signed-off-by: senthil <cendhu@gmail.com>

* rm privacyenabledstate_test pkg and gomega

To add test for the getIndexInfo(), we need to use the
same package name for the test files too. Hence, we replace
the privacyenabledstate_test with privacyenabledstate. Further,
we remove the usage of gomega in the UT.

Signed-off-by: senthil <cendhu@gmail.com>

* add test for getIndexInfo

Signed-off-by: senthil <cendhu@gmail.com>

* Refactor - simplify ledger validation package

Signed-off-by: manish <manish.sethi@gmail.com>

* Doc that Init is not required if you use the contract API

Signed-off-by: NIKHIL E GUPTA <negupta@us.ibm.com>

* Refactor - simplify ledger validation package

Un-export internal structs and functions that were
previously exported because of the code split across sub-packages.

Signed-off-by: manish <manish.sethi@gmail.com>

* Update enable_cc_lifecycle.md

While editing channel ACLs the configurations should be injected using the file enable_lifecycle.json

Signed-off-by: Muhammad Faisal Tariq <mfaisaltariq@gmail.com>

* [FAB-17822]fix incorrect word (#1194)

Signed-off-by: jilg <liguo_ji@outlook.com>

* Add link to Go contract API to smart contract processing topic

Signed-off-by: NIKHIL E GUPTA <negupta@us.ibm.com>

* Update old link in command wrappers

Signed-off-by: NIKHIL E GUPTA <negupta@us.ibm.com>

* Minor fixes to commercial paper tutorial doc

Signed-off-by: David Enyeart <enyeart@us.ibm.com>

* Remove an internal interface in from block storage

Block storage code was initially written such that a more efficient
custom index could be written, given the property that the writes to
this index are append-only. However, for now, we can remove the
interface for a better code navigation.

Signed-off-by: manish <manish.sethi@gmail.com>

Co-authored-by: wuyingjun <wuyingjun@cmss.chinamobile.com>
Co-authored-by: yacovm <yacovm@il.ibm.com>
Co-authored-by: Senthil Nathan N <cendhu@users.noreply.github.com>
Co-authored-by: manish <manish.sethi@gmail.com>
Co-authored-by: NIKHIL E GUPTA <negupta@us.ibm.com>
Co-authored-by: Muhammad Faisal Tariq <mfaisaltariq@gmail.com>
Co-authored-by: Eggsyz <42371983+Eggsyz@users.noreply.github.com>
Co-authored-by: David Enyeart <enyeart@us.ibm.com>