Update master (#34)
* FAB-17910 Ch.Pat.API: join member w/o onboarding (#1431)

This commit handles the following scenario:
* The orderer is a member of the cluster - the config block given at
  join has the self ID of the orderer.
* The config block given at join is number 0 - no onboarding needed.

In this case the join block is first appended to the ledger as a
genesis block, and the multichannel.ChainSupport is created in the
normal fashion, reading the last-block from the ledger and constructing
the consensus.Chain accordingly.

In addition this commit prepares the ground for joining with a join block
with number >0, which requires on-boarding. For that purpose a new
method is added to the consensus.Consenter interface.

This commit also handles the case of an orderer rejecting a join to
the system channel when there are already application channels (FAB-17913).

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I0b2bd4e4da901c6b9c53585bffc6dc0dd5573948

* [FAB-17993] Skip modifying _users security object (#1430)

Introduces a check to skip modifying the security object
of the `_users` table. Future versions of CouchDB
protect from this operation being performed and
it isn't necessary for the peer to actually modify
the object to perform its functionality

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>

* Set Cache-Control no-store for channel participation list (#1421)

FAB-17928 #done

Signed-off-by: Will Lahti <wtlahti@us.ibm.com>

* Channel participation api - redirect from base V1 URL (#1433)

FAB-17892 #done

Signed-off-by: Will Lahti <wtlahti@us.ibm.com>

Co-authored-by: Yoav Tock <tock@il.ibm.com>
Co-authored-by: Brett Logan <brett.t.logan@ibm.com>
Co-authored-by: Will Lahti <wtlahti@us.ibm.com>